update references, relocate some events.
